// src/app/calculators/car-economy-calc/models/car-economy.ts
export class CarEconomy {
    carCost: string;
    city: string;
    divingDistanceIntegration: string;
    drivingDays: string;
    drivingDistance: string;
    fuelConsumption: string;
    fuelCost: string;
    fuelType: string;
    insuranceCost: string;
    parkingCost: string;
    publicTransportIntegration: string;
    travelsPerDay: string;
}